Separate auth and user api and add the ability to refresh auth token (/api/auth/refresh/)
