+++ backend/auth/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AuthConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "auth"
+    label = "authentication"
