+++ backend/auth/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
+from .views import signup
+
+urlpatterns = [
+    path("signup/", signup, name="signup"),
+    path("login/", TokenObtainPairView.as_view(), name="login"),
+    path("refresh/", TokenRefreshView.as_view(), name="refresh"),
+]
